// frontend/app/api/data/interfaces.ts
export interface FormData {
    operation: string;
    number: string;
    date: Date | null;
    headquarterSender: string;
    warehouseSender: string;
    headquarterReceiver: string;
    warehouseReceiver: string;
    basis: string;
    basisHeadquarter: string;
    basisDate: Date | null;
    basisNumber: number | null;
    headRAO: string;
    headRAORank: string;
    headRAOName: string;
    headFES: string;
    headFESRank: string;
    headFESName: string;
    clerk: string;
    clerkRank: string;
    clerkName: string;
    warehouseHead: string;
    warehouseHeadRank: string;
    warehouseHeadName: string;
    warehouseStatus: string;
    recipient: string;
    recipientStatus: string;
    recipientRank: string;
    recipientName: string;
};